feat: increase vote count column type
The `upvote_count` and `downvote_count` columns in the `story_comments` table have been changed from `unsignedInteger` to `unsignedBigInteger`.

This change is to accommodate a larger potential number of votes on story comments, ensuring scalability and preventing potential overflow issues with a very high number of votes.

+++ database/migrations/2025_07_10_131652_add_vote_to_story_comments_table.php
@@ -12,11 +12,11 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('story_comments', function (Blueprint $table) {
-            $table->unsignedInteger('upvote_count')
+            $table->unsignedBigInteger('upvote_count')
                 ->default(0)
                 ->after('reply_count')
                 ->index();
-            $table->unsignedInteger('downvote_count')
+            $table->unsignedBigInteger('downvote_count')
                 ->default(0)
                 ->after('upvote_count')
                 ->index();
